refactor: :recycle: Séparation des routes en modules distincts + correction des tests  - Création du dossier routes/ avec profileRoutes.js et studentRoutes.js - Refactorisation de src/index.js pour utiliser les routes modulaires - Correction des routes profile pour utiliser les bonnes colonnes selon le schéma DB - Amélioration de la gestion des contraintes de clé étrangère dans DELETE profile - Correction des tests pour correspondre à la structure réelle des tables - Tous les tests passent (51/51) avec une couverture complète des cas d'usage  Architecture modulaire et robuste, prête pour l'évolution du projet.
